Move database merge functionality to a separate class
